pyproject.toml
[project]
name = "agent-openai-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asttokens>=3.0.0",
    "comm>=0.2.3",
    "google-genai>=1.46.0",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]
